feat: extract IDP module with authentication and user info handling from 3-layer structure into independent libs
