+++ libs/idp/src/index.ts
@@ -1,0 +1,5 @@
+export * from './idp.module';
+export * from './idp.service';
+export * from './dto/idp-login.dto';
+export * from './dto/idp-token-response.dto';
+export * from './dto/idp-userinfo-response.dto';
