+++ libs/prisma/src/index.ts
@@ -1,0 +1,3 @@
+export * from './prisma.module';
+export * from './prisma.service';
+export * from './prisma-errors';
