--- prisma/prisma.module.ts
@@ -1,9 +1,0 @@
-// src/prisma/prisma.module.ts
-import { Module } from '@nestjs/common';
-import { PrismaService } from './prisma.service';
-
-@Module({
-  providers: [PrismaService],
-  exports: [PrismaService], // This is key! It makes the service available to other modules.
-})
-export class PrismaModule {}
