+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { PostsModule } from './posts/posts.module';
-import { PrismaModule } from 'prisma/prisma.module';
+import { PrismaModule } from '@app/prisma';
 import { AuthModule } from './auth/auth.module';
 import { UsersModule } from './users/users.module';
 import { GoogleStrategy } from './auth/strategies/google.strategy';
