+++ src/posts/posts.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { PostsService } from './posts.service';
 import { PostsController } from './posts.controller';
-import { PrismaModule } from 'prisma/prisma.module';
+import { PrismaModule } from '@app/prisma';
 import { PostsRepository } from './posts.repository';
 import { AuthModule } from '../auth/auth.module';
