+++ src/users/users.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersRepository } from './users.repository';
-import { PrismaModule } from 'prisma/prisma.module';
+import { PrismaModule } from '@app/prisma';
 
 @Module({
   imports: [PrismaModule],
